+++ R/utils.R
@@ -109,6 +109,62 @@ normalize <- function(x, axis = -1, order = 2) {
   )
 }
 
+#' Provide a scope with mappings of names to custom objects
+#' 
+#' @param objects Named list of objects
+#' @param expr Expression to evaluate
+#' 
+#' @examples \dontrun{
+#' # define custom metric
+#' sparse_top_k_cat_acc <- function(y_pred, y_true) {
+#'   metric_sparse_top_k_categorical_accuracy(y_pred, y_true, k = 5)
+#' }
+#' 
+#' with_custom_object_scope(c(top_k_acc = sparse_top_k_cat_acc), {
+#' 
+#'   # ...define model...
+#'   
+#'   # compile model (refer to "top_k_acc" by name)
+#'   model %>% compile(
+#'     loss = "binary_crossentropy",
+#'     optimizer = optimizer_nadam(),
+#'     metrics = c("top_k_acc")
+#'   )
+#'
+#'   # save the model
+#'   save_model_hdf5("my_model.h5")
+#
+#'   # loading the model within the custom object scope doesn't
+#'   # require explicitly providing the custom_object
+#'   load_model_hdf5("my_model.h5")
+#' })
+#' }
+#' 
+#' @export
+with_custom_object_scope <- function(objects, expr) {
+  objects <- objects_with_py_function_names(objects)
+  with(keras$utils$custom_object_scope(objects), expr)
+}
+
+
+objects_with_py_function_names <- function(objects) {
+  if (!is.null(objects)) {
+    object_names <- names(objects)
+    if (is.null(object_names))
+      stop("objects must be named", call. = FALSE)
+    objects <- lapply(1:length(objects), function(i) {
+      object <- objects[[i]]
+      if (is.function(object))
+        attr(object, "py_function_name") <- object_names[[i]]
+      object
+    })
+    names(objects) <- object_names
+    objects
+  } else {
+    NULL
+  }
+}
+
 
 #' Keras array object
 #'